Fixed two bugs:
1) "oldName" contained a decoded sequence, and was later compared with "n" which
was encoded already. This means that if a .desktop file has "%" in its "Name=" entry,
the two will never be equal and its filename will automatically be renamed when
someone changes an unrelated property.

2) When "Name=" != filename, oldName is equal to the "Name=" entry which is then
later compared in slotCopyFinished against properties.url().filename, again,
these two will never be the same resulting in the "Name=" entry being
overwritten by the filename.

svn path=/branches/KDE_3_2_BRANCH/kdelibs/; revision=294787