Adding two new options for editing: * replace tabs * remove trailing spaces
The options are in the "Editing" page of the configure dialog.

I added the following to the commands:
* set-remove-trailing-spaces <bool>
* set-remove-trailing-spaces-save <bool>
* set-replace-tabs-save

The meaning of "set-replace-tabs" has changed to set the dynamic option, the
rest should be pretty self explaining.

I added the following document variables:
* remove-trailing-space <boolvalue>
* remove-trailing-space-save <boolvalue>
* replace-tabs-save <boolvalue>

The meaning of "replace-tabs" is silently changed to set the status of the
dynamic option, the rest is again self explaining I hope ;)

These features will be available in Kate 2.3 which will ship with KDE 3.3

Togeather with the earlier changes to the save cleanups internals, this should be
sufficient for closing 74700.

CCMAIL: 74700-done@bugs.kde.org, kwrite-devel@kde.org

svn path=/trunk/kdelibs/; revision=300331