Patch from Jan Villat that introcuce the concept of "dynamic contexts" into the kate highligting system.
This allows to use cpatured text from a regular expression in rules in a context
created by that regular expression, by marking the context and rules as dynamic
and clooning the contexts with expanded macros in the rules when it is needed.

This allows us to corrctly parse concepts like HERE strings and perls quoting
operators (q[qwrx]//).

Refer to http://bugs.kde.org/show_bug.cgi?id=63289 for the discussions about the
patch.

The patch also contains the required changes to the language.DTD, and a modified
php.xml file that supports HERE strings.

CCMAIL: 63289-done@bugs.kde.org, kwrite-devel@kde.org

svn path=/trunk/kdelibs/; revision=313512