More threading stuff -- now doing directory completion in a separate thread. This has typically been very problematic with slow or stale NFS file systems where ::stat can block for several minutes.
Because of some weird focus stealing stuff that was happens at the same time
this can block the entire desktop for several minutes while waiting to stat
a directory.

This also makes things a lot cleaner -- DirLister was full of hacks to get
around its syncronous nature.  This removes a lot of cruft (150 less lines
of code) and improves performance.

svn path=/trunk/kdelibs/; revision=320911