fix bug 57360: URLs with anchor - sometimes the position is not moved to the anchor
In KHTMLPart::checkCompleted a call to gotoAnchor is performed, however AFAIK
this is useless since at the moment of this call there is no guarantee that
the layouting is done. Result: KHTML jumps to an arbitrary position since
HTMLElementImpl::getUpperLeftCorner(x, y) returns random data (usually it
results in jumping to (0,0) though).

Solution: Create a new (private) slot in KHTMLPart, gotoAnchor. This slots
jumps to the anchor if a HTML reference is present in the current URL. This
slot is connected to the KHTMLView signal finishedLayout(), which is emitted
at the very end of KHTMLView::layout(). When the slot gotoAnchor is invoked,
all HTML elements have well defined positions.

About the "whether or not to jump to the anchor after user interaction"
discussion: KHTML will jump to the anchor even after user interaction. In the
end this is the best behavior, since this is what the user might expect
(after all the HTML reference is clearly visible in the location bar). Also,
unsuspecting users will probably be confused when KHTML does not jump to the
anchor when they scroll the view a bit.

svn path=/trunk/kdelibs/; revision=321509