intercept mouse and keyboard events of form widgets and pass them through the DOM.
This worked before with simple widgets, but failed
for complex widgets, i.e. widgets that contain sub-widgets
and have an eventfilter installed on them.

This also changes the focus policy of widgets to NoFocus,
in order to let KHTML manage the focus passing and prevent
form widgets inside KHTML from disturbing the tabbing chain.

Also removed the two hack classes "EventPropagator" and "FocusHandleWidget".

svn path=/trunk/kdelibs/; revision=333884