Fixing a bug whereby buffer blocks larger than 'bufsize' wouldn't get sent and would make the code enter an infinite loop. Since this code was copied from kextsock.cpp, it's also very likely it contains the same bug over there.
This patch also sets the buffer size to 1460 bytes, which is the maximum
payload of a TCP/IP packet.

I will commit the same fix to KDE_3_3_BRANCH. This is probably not a
showstopper bug, so I am not asking for retagging. However, it is the
RD's prerrogative to include it if so he feels.

CCMAIL:kde-core-devel@kde.org,86975-done@bugs.kde.org

svn path=/trunk/kdelibs/; revision=338142