Fixing a bug reported first in Bug #85851: applications would crash sometimes. I tracked this down to a KResolver object be cancelled, then deleted while still running.
The cancellation would mark the object as KResolver::Cancel, which would cause
the deletion to skip over because it wasn't actually in the Running state.

This has now simplified the code by moving the guardianship over KResolver's
d pointer back to KResolver, as well as the signal emission.

Since this is a large change, I need more testing to make sure the bug
is fixed and that none other have crept in. I also need leak checks. When
this is proven stable, I'll backport to branch and libkopete-compat will
require an update as well.

CCMAIL:kopete-devel@kde.org

svn path=/trunk/kdelibs/; revision=339684