Fix for Craig's UDS_URL commit: allow urls that are not like <current-directory>/<filename>.
Added UDS_FILENAME which is used *only* when moving/copying to another directory,
to figure out which filename to use in the destination directory. The previous behavior
(filename of UDS_URL if set, otherwise UDS_NAME) is kept. This allows kioslaves
to hide internal stuff in URLs and avoid it being used when copying a file.
Thanks to Craig for the idea of the fix.

CCMAIL: craig@kde.org

svn path=/trunk/kdelibs/; revision=344738