Forward port of:
CVS commit by tilladam:

Return 0 instead of a new instance in hold for the data protocol. Otherwise
kill() doesn't work since the new instance is not connected to any slots
and it's not needed anyhow. David came up with this, I just ran into the
problem in the vcard attachment formatter in KMail.

svn path=/trunk/kdelibs/; revision=344956