New outline painting algorithm.
While the 3.2 outline painting algorithm was quite an improvement over the old
one, it still had issues when elements of the line exceeded the line height.
The new algorithm fixes those finally.

It furthermore coalesces the actual drawing operations to a minimum, which makes
it more efficient under X11, and also ensures that straights are painted
in one go (especially noticeable for styles dotted, and dashed).

Known issues are refresh problems of the outline on focussed links: Only the
part covering the line height is regarded. This is a different bug that has
to be investigated separately.

svn path=/trunk/kdelibs/; revision=347042