Eliminating flicker in KPlotWidget.  Added "WNoAutoErase" flag to QWidget ctor.  It works well!
(note: I first tried to add "setBackgroundMode( QWidget::NoBackground )"
to the KPlotWidget ctor, but this didn't stop the flicker.  This is
strange, because that's how I got rid of the flicker in the SkyMap
widget.  SkyMap doesn't use WNoAutoErase, yet it is flicker-free.  Odd.)

CCMAIL: kstars-devel@kde.org
CCMAIL: cniehaus@kde.org

svn path=/trunk/kdeedu/libkdeedu/kdeeduplot/; revision=357248