Adding support for a new UDS Atom : UDS_LOCAL_PATH
It's used by ForwardingSlaveBase when the corresponding file is available
on the local filesystem (useful in particular for media:/).

I introduced a new accessor to KFileItem if we want to use this atom more.

CopyJob already take care of it (useful when you try to trash a file
from media:/ since only local files can be trashed).

svn path=/trunk/kdelibs/; revision=359824