Use a QListBoxItem's text() and next() instead of indexing through the list box, using text().  QListBoxes use doubly-linked lists internally to store their entries, so using next() is faster than indexing.
No infinite loops today... :)

As an aside:
------------

The difference in speed is not as much as you might think though. QT have been
rather clever in their implementation of QListBox: the pointer to the
QListBoxItem for the last requested index is cached; then, when a different
index is asked for they just apply next() or prev() the relevant number of
times to that pointer and return the entry.  So if you're incrementally going
through the list, it only has to do a few comparisons to see it needs to do
next() once then cache the new QListBoxItem pointer and return the entry.

svn path=/trunk/kdelibs/; revision=362424