Make as-you-type spell checking work with aspell.
Also replace some QTimer::singleShot calls with a standard QTimer, this
probably won't change the semantics of the code but I'm paranoid about
a race condition causing the slot associated with the timer to be called
multiple times (erroneously).

I checked to make sure that I didn't break normal spell checking.

Basically the problem is that aspell has a habit of emitting spurious
white space to stdout. This breaks the inter process
communication (IPC) between kspell.cpp and the spell process.

The IPC between kspell.cpp and the spell process is fragile. kspell.cpp
uses a pair of macros OUTPUT and NOOUTPUT to hook and unhook the
readReady signal from the spell process to a slot in kspell.cpp. This is
fragile because if the spell process emits data to stdout when the
readReady signal is not hooked up to any slot (which happens due to
the spurious emissions by aspell) then the KProcIO instance
managing the spell process will emit a readyRead signal and then not
emit a further readyRead signal until the data that is ready for reading
is read by kspell.cpp. Unfortunately due to the fragile logic of kspell.cpp
this will never happen (because the emission occurred outside of a
OUTPUT/NOOUTPUT connect/disconnect pair).

Thus before kspell.cpp pipes a word into the stdin of the spell process for
checking it must flush any data that has been written to stdout by the
spell process and buffered by KProcIO.

svn path=/trunk/kdelibs/; revision=366042