BACKPORT:
fix memory leak.
mLicenseText is allocated using qstrdup() in setLicenseTextFile().

From QT doc:
char * qstrdup ( const char * src )
Returns a duplicate string.
Allocates space for a copy of src, copies it, and returns a pointer to the copy. If src is 0, it immediately returns 0.
The returned string must be deleted using delete[].

svn path=/branches/KDE_3_3_BRANCH/kdelibs/; revision=368380