HAPPY CHRISTMAS ALL!  My present to you all - My biggest patch yet.
When you select text and copy it to clibboard, it now retains the formatting.
It's fairly complex, so there will possibly be bugs despite my testing.
I'd really like any feedback.

It does not deal with css, so the formatting may be a little off.  I do not think I will 'fix' this as I don't see it being that broken a behaviour.  (feel free to persuade me otherwise)

It is guaranteed to produced valid xml.

When copying text+images, the images are pasted as urls (since I'm using mimetype text/html).  I'm looking at perhaps embedding the images in the clipboard using a webarchive.

You can right click an image and copying the image.  This now actually copies it to the clipboard as well.  This shows up bugs in openoffice (not dealing with transparency correctly for some reason) and kword (it prefers to paste as text/plain over image/* )
Also note that a copied image is the image you see.  Not the image downloaded.

Please note that no functionality has been removed.  Copying and pasting text to a program that does not support rich text, pastes it without formatting, exactly as it did before.  For apps that do support rich text, I strongly recommend all such apps implement 'paste special' of some sort so the user can chose to paste without formatting.  I may work on this at some point.

You can test this by pasting into openoffice.  Support for pasting into kword will come soon I hope (I need to get kword to run it's import filters on the clipboard data)



FEATURE:89057
FEATURE:36029
FEATURECC:71476

svn path=/trunk/kdelibs/; revision=372574