Make the selectedTextAsHtml  work again.  It's not perfect, but it doesn't crash, so it's good enough for testing :)
Leo, could you have a look at this please.  When the end (or start) of the selection is an image, it uses the parentnode and calculates the offset, but for some reason it's doing it wrong.  I use the konqueror start screen for testing.

It works fine otherwise.

CCMAIL:Leo Savernik <l.savernik@aon.at>

svn path=/trunk/kdelibs/; revision=376996