This completely re-written WhatsThis explains purpose and usage of the 'Additional Tags' tab widget. This widget is rarely used, but were it is, has proven to be very useful for printing in enterprise enviroments.
However, it was also partly broken, in that it sent the print option with
a 'KDEprint-' prefix in the name (requiring the downstream filter to be
modified so it recognizes that prefix). The new behviour is to only use
the 'KDEprint-" prefix internally, and strip it off when the job is sent
downstream.

That means we can now support arbitrary (CUPS- and Third-Party-)joboptions
which may be developed in the future, or which are currently unsupported
by KDEPrint (they are rare, but there are a few...). [One such usage is for
customized banner pages, containing all kinds of additional info, such as
the Room Number or the Phone Number of the job owner, where this info is
transported through the use of variables filled in on the 'Additonal Tags'
tab.]

Unfortunately the settings currently cannot be saved across kprinter starts.

The new WhatsThis explains that in detail.

This is a partial fix for bug # 56163.

BUG:56163

svn path=/trunk/kdelibs/; revision=384172