This is a strange file.
The <language ...> is missing extensions="..." AND mimetypes...
So this highlighting is never used automatically.

The item datas used defaultStyles that did not exists (like dsAssignment or
dsMatrix). I changed this to dsOthers and dsNormal.

Maybe someone can enlighten us about extensions, now the file should even work.

I send this to the one who added this file to CVS at 2004-06-16 as well:
CCMAIL: bramschoenmakers@kde.nl

svn path=/trunk/kdelibs/; revision=389182