Fixed loss of toolbar settings when saving settings after switching between parts. Let me explain how it happened:
- toolbar is created
- loadState is called to load the app defaults from the XML, which are stored into FooDefault member vars
- applySettings is called to load user settings from KConfig

switching to another part:
- saveState is called to save the current settings to XML (in memory)
- the toolbar can be deleted then

switching back to this part:
- the toolbar is recreated
- loadState is called to load the settings from XML - those are not the application defaults anymore!
- applySettings is called to load user settings from KConfig

Then when saving settings: "reverting to default" is done wrongly,
since not comparing with the real application default, but with the current settings
(so everything appeared to be equal to the app defaults, and all settings were deleted).

This is fixed by saving the FooDefault member vars into the XML, so that when
loadState is called the 2nd time, all those values are restored correctly.
Danimo will close the relevant kontact bugs after re-testing.

svn path=/trunk/kdelibs/; revision=391794