Ok, this was one of our "showstoppers" here at SAP (combined with a kio_ldap commit that will follow).
KPasswdServer::checkAuthInfo() wasn't checking to see if the auth info was
available in KWallet.  The implications of this were that it was impossible
to have a slave (LDAP specifically) check for authentication credentials
without popping up a dialog.  In practice this meant that every time the LDAP
slave was restarted (i.e. after a timeout) the user would be prompted for the
password again.  When used with KMail integration (as is fairly essential in
this environment) this meant prompting the user for their LDAP password pretty
much every time they used it to send a new mail.

And with that fine over-explanation, here we go.  ;-)

svn path=/trunk/kdelibs/; revision=392296