This patch does the following - implement splitterKeepSize correctly. - Internally use always highResolution due to *ugly* rounding bugs. - in fact, the high resolution mode never worked 100%. KDockWidget   always assumes a splitter position between 0..100 but it was 0..10000,   which lead to size-bugs undock/redock actions.
For KDE 4, remove this highResolution option completely.
This patch should be BC, kdockwidget_private.* is private API.

Compatibility: The splitter position must be always between 0..100.
If it is >100, it assumes high resolution from the old behavior, and
it will be fixed to a correct value between 0..100. This makes sure
it works for both, high resolution and not high resolution.

I tested this with k3b, kate, konqueror and a bit Quanta.
If you find bugs, please let us know on kmdi-devel@kde.org.

This fixes the bugs
- #69895: selection menu gets not properly positioned on the right side
- #75112: borders in IDEAl mode are not "active"

CCBUG:69895
CCBUG:75112

CCMAIL:kmdi-devel@kde.org

svn path=/trunk/kdelibs/; revision=394730