Always get the current url of the job: added the new redirectionURL() method to KIO::ListJob and made KDirListerCache check for it whenever a job's url is needed. Except for the redirection case---here we really want the old url.
I've still got some problems with kio_lan crashing all the time, which makes
debugging this KDirListerCache case pretty hard... and I'm not sure if
it's worth the effort that KDirLister continues to work flawlessly even if a
job crashed? Will try to fix that tomorrow, debugging it takes a hell lot of
time.

svn path=/trunk/kdelibs/; revision=400133