Instead of returning the original domain lowercased, rejoin the multiple parts with periods. In most cases, this won't make a difference, but IDN spec allows other separators in domain names other than the dot.
I do not believe this to be a security risk, nor a source of
phishing/scams. Unlike the homograph attack (which prompted this
return to be added in the first place), you cannot create a
near-looking domain with this feature: the domain IS the same.
i.e., www.google.com === www.googleãcom

CCMAIL:security@kde.org

svn path=/trunk/kdelibs/; revision=403231