Fixing the random resolver failures in the code. It was a local error after all, so I apologise for being hard on the bug reporters. You know how developers are protective of their own code :-)
Many thanks to the patient bug reporters and to Coolo for his analysis
of the problem.
BUG:94703

The reason this bug happened was quite insidious. It was related to
some events occurring in a very particular order in different threads,
that's why it appeared to be random.
- the lookups are started (KResolver::start())
- KResolver::wait() is called on the master thread
- the lookups finish on the auxiliary threads
- the resolver code detects the auxiliary lookups being done and
  processes the results (KResolverManager::doNotifying()), thereby
  waking up all threads on KResolver::wait()
- the master thread is woken up now
- here's the catch: while the master thread is waking up, the manager
  thread has started processing the final results
  (KResolverManager::handleFinishedItem()) and sets status to
  KResolver::Success
- the master thread thinks the resolving is done and emits the
  finished(...) signal with an empty KResolverResult list!
- after that, the manager thread collects the auxiliary results,
  builds the main results and emits the signal again, but it's too late,
  since an error will have already been reported

After understanding the error, I am actually surprised it hasn't
happened more often, least of all with me. I am betting it's the
different threading implementations that cause the different
behaviour, or the fact that people were using dual-processor or
dual-core systems (which can do threading better than my single-core
CPU).

svn path=/trunk/kdelibs/; revision=404040