Backport: SVN commit 410734 by mbuesch:
fix memory leak.
From QT documentation:

bool QApplication::sendEvent ( QObject * receiver, QEvent * event ) [static]
<SNIP>
The event is not deleted when the event has been sent. The normal approach is to create the event on the stack, e.g.
 QMouseEvent me( QEvent::MouseButtonPress, pos, 0, 0 );
 QApplication::sendEvent( mainWindow, &me );
If you create the event on the heap you must delete it.

svn path=/branches/KDE/3.4/kdelibs/; revision=410736