The evening begins quite well -- that was easy to fix, though a bit embarrassing :-\ My redirection code wasn't perfect.. ehm, complete, I forgot about 2 important lines.
What happened is this: When pressing up while being in an archive we end up in 
e.g. tar:/home/michael, which redirects to file:///home/michael. The item is in
the cache already, but in this case I forgot to insert the listers and holders
with the new url back into the KDirListerCache. Sorry...

Many thanks for all the reports! Much appreciated.

BUG: 103795, 105827, 104973

svn path=/trunk/KDE/kdelibs/; revision=415093