Introduce a new property for desktop files : X-KDE-Protocols
Allow to declare which protocols are really supported for special cases
(in particular applications needing to seek in files).

The behavior is the following:
1) If X-KDE-Protocols is not specified or X-KDE-Protocols=KIO,
   %U works as expected
2) If X-KDE-Protocols=protocol1,protocol2,...
   %U works as expected for protocol1, protocol2, ...
   Other URLs are rewritten using NetAccess::mostLocalURL() before
   being passed to the application (to ensure having file:/ URLs when using
   media:/ or system:/ for example)

As a result it's not necessary to resolve URLs having UDS_LOCAL_PATH
in KFileItem::run(), everything is processed in KRun.

Thanks a lot to dfaure for helping with KRun itself.

svn path=/trunk/KDE/kdelibs/; revision=420946