Backport of:
SVN commit 424591 by tilladam:

Don't reparse the full KGlobal configuration (which takes around 2
seconds here) each time a KDictSpellingHighlighter is instantiated,
which is everytime a KMail composer is opened. The only key that is
read there is a hidden, undocumented, non-gui-accesible tuning key,
so there is zero probability of it changing under us by any means
but direct editing of the config file. In that case the user can
reasonably be expected to restart the application, especially if it
means that starting a composer KMail becomes instantaneous again,
for all users.

Ok'd by Waldo and Zack.

svn path=/branches/KDE/3.4/kdelibs/; revision=424606