Crash--;
The linerange class needs to be passed as pointers because there's some
interaction with copying of the ranges and deletion of the layouts.
However, several functions in the view return new lineranges which are
not part of the cache, so we'll have to track them and delete when
they're no longer in use.

Hopefully I can find the reason for the strange deletion first, that's
easier at the moment.

Still has a crash while scrolling the view...

svn path=/branches/work/kde4/kdelibs/; revision=425764