Major patches for KDE autoconf scripts.
These patches provide the KDE configure scripts (generated by the configure.in.in files spread all 
over the KDE modules) with --with-<option> and --without-<option> switches in many places where 
these switches are missing. This gives the user that compiles KDE the possibility to override the 
autodetection method used in these scripts to decide whether to make use or not of an external 
library or tool. 

Having the possibility to override all the autodection of external software has a very concrete 
benefit: it makes it possible for KDE packagers to build a customized version of a KDE package 
independently of the building host; and for source-based distributions, it makes it possible to 
define the dependency tree of the package in a deterministic way.

The coding style of these patches is the one recommended by autoconf developers, as illustrated in 
the info page of the AC_ARG_WITH macro.

See http://dev.gentoo.org/~greg_g/kde-configure/ for more information.



svn path=/trunk/KDE/kdelibs/; revision=433982