Merging branch in branches/work/kdecorenetwork-srv-resolution back to the main line: implements SRV-based DNS lookups according to RFC 2782.
The code is in working condition and should compile for
everyone. Bug me if it doesn't.

Missing feature: sort the per-priority entries according to weight. This
is a "SHOULD" feature according to the RFC. I'll implement it somewhen.

Foward-port to KDE4 will not happen now because there's no QDns replacement
outside the Qt3Support library.

svn path=/branches/KDE/3.5/kdelibs/; revision=443592