The Help->About boxes use these license files.  This means that the help boxes had the text:
This General Public License does not permit incorporating your program into
proprietary programs.  If your program is a subroutine library, you may
consider it more useful to permit linking proprietary applications with the
library.  If this is what you want to do, use the GNU Library General
Public License instead of this License.

This seems just plain wrong to have in the license text in the about box for a program!  So I've removed this text.


svn path=/trunk/KDE/kdelibs/; revision=446184