Fixed slow directory update in KFileDialog and Konqueror: the KFileItem comparison was determining the mimetype of both kfileitems, which is a very slow operation when they have no known extension: reading the contents is then necessary.
There's no reason to compare the mimetypes, in fact: it's pretty common that the view
(kfiledialog or konq) has determined the mimetype after the listing, and that the ioslave
doesn't provide a mimetype during listing. Also it's pretty hard to change
the mimetype of a file without changing either its name, size, or modification
time.

svn path=/trunk/KDE/kdelibs/; revision=456564