when the year selector was open, if you clicked on any other part of the calendar aside from the year toggle button the inner event loop started by kframepopup would never get exited. this resulted in the calendar never actually deleting it, and resulting in bugs like #112207.
now the event loop is exited properly, deletions happen and everything
behaves as expected. the single shot timer to check the state of the
button is there because the button must be manually raised when the
popup goes away do a click elsewhere but must NOT be raised manually
when the button is clicked (otherwise you get a false press and the year
popup reappears instantly). however, the button's "down" property only
changes once we go back to the event loop, ergo the timer. the actual
button is clicked signal happens after that, interestingly.

anyways. nasty bug. =(

BUG:112207

svn path=/branches/KDE/3.5/kdelibs/; revision=464357