Fix this not to corrupt memory so kdelibs build can run kconfig_compiler w/o aborting w/a heap corruption warning from glibc.
Please, please, please. If you are going to do the "move the data to d" thing 
for a class instances of which can be copied you implement both the copy constructor 
and the assignment operator. And if the class can't be copied, make sure those are private.
(QObject subclasses get that for free, and see Q_DISABLE_COPY).

And, please^{*}: do not use stuff like this:
#define mAppName d->_appName
as that's just unreadable and confusing. 

CCMAIL:brax108@cox.net

svn path=/trunk/KDE/kdelibs/; revision=466242