CCMAIL: groot@kde.org CCMAIL: lists@stevello.free-online.co.uk DOX: Improving apidox of KIMIface and its companion KIMProxy
This is my first go on improving the 3.5 branch apidox, hope I got
it right :)

Added information to the KIMIface apidox mainly targetted at developers
who want to implement it based on implementing it myself and on clarifications
by Will while doing that.

Syncronized the apidox of KIMProxy with the additions in kimiface.h and
added a short example code snippet to the class docs.

Both classes: replaced @ref with methodname() as recommended in Adriaan's
apidox howto/style guide.

Added explicit @brief descriptions to each method, but I see that
BRIEF_MEMBER_DESC is deactivated in our Doxygen.global file.
Actually I find that short description in the member overview quite useful,
maybe we can activate that. Adriaan?

Modified the Makefile.am files in both directories to have doxygen pull
references of the DCOP and KABC submodules.
The weird thing is that if I add "interfaces" to the doxygen variable it
nicely references the other class (KIMIface from the KIMProxy docs and vice versa)
but it then failes to add the KABC references. *sigh*

Will: I changed the documentation of KIMProxy::imAppsAvailable() because as
far as I understood the KIMProxy internals the list of KIMIface stubs is
actually the list of running KIMIface applications, not the list of just
installed ones.


svn path=/branches/KDE/3.5/kdelibs/; revision=466512