ok, weird.
I had mark partial files turned of, because, uhm, it's stupid.  Well, apparently
it got confused and would name it .part anyway, but since mark partial files was
turned off, it'd not rename it away from ".part"

So, the solution should be all of the following:
1) We eliminate the scourge that is .part files (and replace it with a better
metadata solution)
2) when you download files from ftp, it should still use kio_file for writing
the files locally.


svn path=/branches/KDE/3.5/kdelibs/; revision=468896