Generate Job.moc and Thread.moc manually so that the QObject derived classes that are defined in the .cpp files are included in the appropriate moc files
Add a file that includes the .moc files for the other classes that need it. I
chose to do it this way so as to modify as little of the original source as 
possible, since this is supposed to build with both qmake and unsermake.

Idea stolen from Kopete's jabber plugin where we had to do the same thing.


svn path=/trunk/kdenonbeta/threadweaver/; revision=488605