HiColor is moved to KDEArtwork.
Except for the directory and the "index.theme" file.

It is made visable so the the user can choose it -- doesn't make much difference (except for GNOME icons [if installed]) till it is fully supported.

Icons to be installed as HiColor should be in KDEArtwork.

svn path=/branches/KDE/3.5/kdelibs/; revision=489334

The following changes were in SVN, but were removed from git:

M	pics/hicolor/Makefile.am
M	pics/hicolor/index.theme