Typedef QFlags<ButtonCode> to ButtonCodes
According to the Qt guidelines:
http://doc.trolltech.com/4.0/qflags.html

Quote:
A sensible naming convension for enum types and associated QFlags types is to give a singular name to the enum type (e.g., Option) and a plural name to the 
QFlags type (e.g., Options). When a singular name is desired for the QFlags type (e.g., Alignment), you can use Flag as the suffix for the enum type (e.g., 
AlignmentFlag).


svn path=/trunk/KDE/kdelibs/; revision=496608