KDialog centering
Since KDE Kicker (and win32's "kicker") is "always on top", 
it's better for centered dialogs to use availableGeometry() 
instead of screenGeometry(), 
to avoid covering them by Kicker (it's a problem for default KDE settings).

Also affects KDialog::avoidArea().

As QDesktopWidget::availableGeometry() is a subrect 
of QDesktopWidget::screenGeometry(), 
the fix will not affect systems with many displays.


svn path=/branches/KDE/3.5/kdelibs/; revision=499587