Fix the extraction of two-parameter tr keyword in Qt.
(The first parameter is the text, the second one is a comment, which KDE4 can
use as context. KDE3 did the other way round but that does not work with Qt4.)

Unfortunately it means that in KDE4 the translation for left-to-right
or right-to-left in Qt will not be the same message than in Qt3. :-(

CCMAIL:kde-i18n-doc@kde.org
CCMAIL:kde-core-devel@kde.org


svn path=/trunk/KDE/kdelibs/; revision=506075