Remove the lines in the Doxygen comment of setTranslators telling that the translators' placeholders are added automatically by Scripty.
For doing this, Scripty has to grep all C/C++ source and header files
and e.g. Kopete has still not automated placeholders.

I am not making the placeholder manadatory in code yet but developers can
start to care about them a little more.
(As it was already the case during KDE3, when the KAboutData class reference 
was not in the same directory than the Makefile.am with the "messages" target.)

svn path=/trunk/KDE/kdelibs/; revision=506087