This commit add support for signals in kconfig_compiler. What does it means ? It means that a changed entry can emit a signal to dispatch it to another component.  I needed it in Kopete to replace a old hand-made configuration class.
First you need to define a list of signals in the configuration file like 
this.
<kcfg>
<signal name="emoticonSettingsChanged" />

<signal name="styleChanged">
  <label>Tell when a complete style change.</label>
    <argument type="String">stylePath</argument>
    <argument type="String">StyleCSSVariant</argument>
</signal>
</kcfg>

After, add to the entry the signal to emit:
<entry key="stylePath" type="String">
  <label>Absolute path to a directory containing a Adium/Kopete chat window 
style.</label>
  <emit signal="styleChanged" />
</entry>

To acheive it, I needed to make KConfigSkeleton derive from QObject to be able 
to call signals in derived classes. 
I define a enum list in the devirated class that I store in a bitmask to 
determine which signal was modified. Then in the derivated usrWriteConfig() 
method, it emit the changed signals. 

Also KConfigSkelton has a new signal called "configChanged()" when to tell when the configuration file was updated.

Commit include also the new test files, modifications to build system (automake and CMake), addition to the 
Doxygen files.


svn path=/trunk/KDE/kdelibs/; revision=506883