Fix a very subtle bug (the analysis to which was under my nose all the time): when a document gets detached from inside an event handler of a widget, it detaches the tree, destroys the renderer arena and sets a new one. The problem is, the RenderWidget doesn't actually get detached, as it's held by a protecting refcount. The end result is that when the deref() is called on the RenderWidget, it tries to free itself... from the new arena. Hence, memory corruption, crashes, etc. Except, of course, it's within the arena, so valgrind doesn't see anything. And if you build without the arena? A very truthful assert saying that the arena pointer got changed, somehow --- which yours truly interpreted as memory corruption, and was bewildered by the lack of vg warnings on it.
My fix is to have the RenderWidget a co-owner of the arena, with its own pointer,
and ability to prolong its life. 

Oh, and if any of the build system gods is still reading, I would be eternally grateful 
if someone added a configure check for Valgrind headers, so I can instrument this
damn arena thing with VG hints once and for all, and save all of us bajillions of times
debugging render tree messes.

BUG:121513



svn path=/branches/KDE/3.5/kdelibs/; revision=508195