Reverting - generating moc files in subdirs worked, but this broke compilation because the subdir (in the builddir) is missing from the -I. How could we add it?
(E.g. kdelibs/kio/kfile/kopenwith.cpp is compiled from kdelibs/kio, so the relative path
to the header is kfile/kopenwith.h, and this created a kfile/kopenwith.moc, but it's included
as "kopenwith.moc" from kopenwith.h so we need -Ikfile in the compilation line. This is
necessary for the kabc/addressbook.h vs kabc/kab/addressbook.h case)

CCMAIL: kde-buildsystem@kde.org

svn path=/trunk/KDE/kdelibs/; revision=509646