Check if the redirection is for the same URL and give up if it is the case.
Note: I am commiting this now, as Debian bug #355111 seems to indicate 
a uncooperative behaviour toward the server (with a tendency to be a denial of
service). However I am not sure about the patch, as the URLs could differ
only by trailing slahes. I do not know if this would be valid redirection.
That is why I am not closing the bug either, to let an expert decide.
(If it is not valid, then KURL::equals should be used instead of
KURL::operator == )

CCBUG:123034

svn path=/branches/KDE/3.5/kdelibs/; revision=515665