* link not accessable with negative text-indent (#96275) * Wrong :hover effect with negative text-indent (#90510) * text not selectable if it has a negative text-indent
If ever you've had text that oddly won't select, or text that won't get a
"text input" cursor when you hover over it. Or Links that you can't 
click on, then this is the patch for you.

This patch works by adding "negative overflow" - it is just like a normal
overflow, except it goes on the other side.

Whenever you use effectiveWidth(), you should *always* use effectiveXPos() with
it instead of just xPos().

(no new regressions according to testregression)


svn path=/branches/KDE/3.5/kdelibs/; revision=516017