Don't attempt to (synchronously) save if we are still async loading, because otherwise the following fascinating chain of events is possible:
- Kontact shutdown unloads parts, which destructs the addressbook part
- StdAddressbook is deleted, which clears the addressbook
- before the clear the composer in KMail finds a closed addressbook,
  thus opens it again, triggering an asyncLoad.
- that asyncLoad triggers a KIO::Job, which starts, but never finishes
- KApplication's static deletion of the newly opened addressbook leads
  to a saveAll, which leads to a save() of the file resource, which 
  iterates over the StdAddressbook which has already been cleared, but
  not re-loaded yet, since we are waiting for the KIO copy job of the 
  temp file to finish, thus writing out a 0 byte file

The reason why the kio job doesn't finish is that it has already been
killed, at this point (we are in static deletion of the application,
after all).

I'm suspecting that StdAddressbook should not call saveAll() from the
destructor. A wise man (David Faure) once told me "Intelligent dtors are
bad", and I've since painfully learned a few times how right he was.
I'm afraid to change that in branch, though, no telling what relies
on it. Also KMail's composer auto-save should not hit the KeyResolver,
which loads the addressbook, but that's only marginally related.

This fixed the 0-file issue described in the bug, but it's still writing
out spurious empty entries, sometimes, to the file resource. Will debug
that next.

BUG: 118924
CCMAIL: tokoe@kde.org

svn path=/branches/KDE/3.5/kdelibs/; revision=516587