reinstate some code I removed in r516516 as that was a mistake ;(
the widgets list is used not only to establish a clip region on the painter but also to keep 
RenderWidgets and their associated widgets in sync.

Indeed, as we can only specify an accurate position of widgets when painting, RenderWidget leaving the viewport
would not be able to update their widget to their new position, as they would not be painted anymore...

Added some comment to explain this.


svn path=/branches/KDE/3.5/kdelibs/; revision=520078