fix bug 120605: wikipedia editing preferences render incorrectly
Caused because when you do a CSS "clear", this is occasionally ignored
when calculating the width of a box.

Causes a couple of fake regressions (baseline update soon). One of the
regressions has a new failure which is slightly better than the incorrect
baseline! :)

css21/t0905-c5525-fltmult-00-d-g.html
webcore/fast/block/float/014.html

BUG:120605


svn path=/branches/KDE/3.5/kdelibs/; revision=520803